requirements.txt
adium-theme-ubuntu==0.3.4
chardet==2.3.0
croniter==0.3.8
cryptography==1.2.3
Django==1.10.6
enum34==1.1.2
gitdb==0.6.4
GitPython==1.0.1
idna==2.0
ipaddress==1.0.16
Jinja2==2.8
MarkupSafe==0.23
msgpack-python==0.4.6
ndg-httpsclient==0.4.0
Pillow==3.1.2
pyasn1==0.1.9
pycrypto==2.6.1
pycurl==7.43.0
PyMySQL==0.7.2
pyOpenSSL==0.15.1
python-apt==1.1.0b1
python-dateutil==2.4.2
python-systemd==231
pytz==2014.10
PyYAML==3.11
pyzmq==15.2.0
requests==2.9.1
salt==2015.8.8
simplejson==3.10.0
six==1.10.0
smmap==0.9.0
tornado==4.2.1
unity-lens-photos==1.0
urllib3==1.13.1
vboxapi==1.0
djangorestframework==3.6.3
djangorestframework-xml==1.3.0
